CryptoJournal/requirements.txt
pycryptodome
dropbox
